#2 sqlite read-only mode

+++ storage/sqlite/storage.go
@@ -15,11 +15,15 @@ import (
 type SQLiteStorage struct {
 	DB            *sql.DB
 	activeModules map[string]bool
+	readOnly      bool
 }
 
-func NewStorage(ctx context.Context, path string) (*SQLiteStorage, error) {
+func NewStorage(ctx context.Context, path string, readOnly bool) (*SQLiteStorage, error) {
 
 	if _, err := os.Stat(path); os.IsNotExist(err) {
+		if readOnly {
+			return nil, errors.New("storage does not exist and is read-only")
+		}
 		file, err := os.Create(path)
 		if err != nil {
 			return nil, err
@@ -37,10 +41,14 @@ func NewStorage(ctx context.Context, path string) (*SQLiteStorage, error) {
 		sqliteDatabase.Close()
 	}()
 
-	return &SQLiteStorage{DB: sqliteDatabase, activeModules: make(map[string]bool)}, nil
+	return &SQLiteStorage{DB: sqliteDatabase, activeModules: make(map[string]bool), readOnly: readOnly}, nil
 }
 
 func (s *SQLiteStorage) Write(ctx context.Context, d model.Data) error {
+
+	if s.readOnly {
+		return errors.New("storage is read-only")
+	}
 
 	if ok, err := s.moduleActive(ctx, d.Module); err != nil || !ok {
 		return err
@@ -136,7 +144,7 @@ func (s *SQLiteStorage) moduleActive(ctx context.Context, module string) (bool, 
 		return true, nil
 	}
 
-	if _, ok := s.activeModules[module]; !ok {
+	if _, ok := s.activeModules[module]; !ok && !s.readOnly {
 		q := fmt.Sprintf("CREATE TABLE IF NOT EXISTS `%s` (DateTime TEXT, Topic TEXT, Value TEXT)", module)
 		_, err := s.DB.ExecContext(ctx, q)
 		if err != nil {
@@ -150,6 +158,9 @@ func (s *SQLiteStorage) moduleActive(ctx context.Context, module string) (bool, 
 
 // Cleanup removes the table for the given module
 func (s *SQLiteStorage) Cleanup(module string) {
+	if s.readOnly {
+		return
+	}
 	q := fmt.Sprintf("DROP TABLE `%s`", module)
 	s.DB.Exec(q)
 	delete(s.activeModules, module)
